+++ metadoc/fastmath/core_examples.clj
@@ -348,13 +348,19 @@
     (rank [3 1 4 1 5 9 2 6 5 3 5] :random)))
 
 (def single-list `(sin cos tan cot sec csc asin acos atan acot asec acsc haversine
-                       sinh cosh tanh coth sech csch asinh acosh atanh acoth asech acsch
-                       qsin qcos exp log log10 ln log1p expm1 sqrt cbrt qexp qsqrt rqsqrt
-                       erf erfc inv-erf inv-erfc sinc log2 qlog log1pexp
-                       sq cb pow2 pow3 safe-sqrt floor ceil round rint abs iabs trunc itrunc
-                       frac sfrac low-2-exp high-2-exp round-up-pow2
-                       signum sgn sigmoid logit
-                       gamma log-gamma digamma log-gamma-1p trigamma inv-gamma-1pm1))
+                     sinh cosh tanh coth sech csch asinh acosh atanh acoth asech acsch
+                     qsin qcos exp log log10 ln log1p expm1 sqrt cbrt qexp qsqrt rqsqrt
+                     erf erfc inv-erf inv-erfc sinc log2 qlog log1pexp
+                     sq cb pow2 pow3 safe-sqrt floor ceil round rint abs iabs trunc itrunc
+                     frac sfrac low-2-exp high-2-exp round-up-pow2
+                     signum sgn sigmoid logit
+                     gamma log-gamma digamma log-gamma-1p trigamma inv-gamma-1pm1))
+
+(add-examples floor
+  (example-session "Scaled floor" {:test-values [230.0 234.3]} (floor 234.312 10) (floor 234.312 0.1)))
+
+(add-examples ceil
+  (example-session "Scaled ceil" {:test-values [240.0 234.4]} (ceil 234.312 10) (ceil 234.312 0.1)))
 
 (def double-list `(bessel-j atan2 hypot hypot-sqrt log-beta))